Move address controls into shared component. This should prevent flakey tests due to IDs changing
