+++ plugins/woocommerce-blocks/assets/js/blocks/checkout/edit.tsx
@@ -1,7 +1,6 @@
 /**
  * External dependencies
  */
-import { __ } from '@wordpress/i18n';
 import clsx from 'clsx';
 import {
 	InnerBlocks,
@@ -14,10 +13,9 @@ import {
 	previewCart,
 	previewSavedPaymentMethods,
 } from '@woocommerce/resource-previews';
-import { PanelBody, ToggleControl, RadioControl } from '@wordpress/components';
 import { SlotFillProvider } from '@woocommerce/blocks-checkout';
 import type { TemplateArray } from '@wordpress/blocks';
-import { useEffect, useRef, useState } from '@wordpress/element';
+import { useEffect, useRef } from '@wordpress/element';
 import { getQueryArg } from '@wordpress/url';
 import { dispatch, select as selectData, useSelect } from '@wordpress/data';
 import { store as coreStore } from '@wordpress/core-data';
@@ -34,7 +32,7 @@ import {
 	useBlockPropsWithLocking,
 } from '../cart-checkout-shared';
 import '../cart-checkout-shared/sidebar-notices';
-import { CheckoutBlockContext, CheckoutBlockControlsContext } from './context';
+import { CheckoutBlockContext } from './context';
 import type { Attributes } from './types';
 
 // This is adds a class to body to signal if the selected block is locked
@@ -100,52 +98,6 @@ export const Edit = ( {
 		};
 	} );
 
-	// These state objects are in place so the required toggles remember their last value after being hidden and shown.
-	const [ lastPhoneRequired, setLastPhoneRequired ] = useState(
-		defaultFields.phone.required
-	);
-	const [ lastCompanyRequired, setLastCompanyRequired ] = useState(
-		defaultFields.company.required
-	);
-	const [ lastAddress2Required, setLastAddress2Required ] = useState(
-		defaultFields.address_2.required
-	);
-
-	useEffect( () => {
-		if ( ! defaultFields.phone.hidden ) {
-			setLastPhoneRequired( defaultFields.phone.required );
-		}
-		if ( ! defaultFields.company.hidden ) {
-			setLastCompanyRequired( defaultFields.company.required );
-		}
-		if ( ! defaultFields.address_2.hidden ) {
-			setLastAddress2Required( defaultFields.address_2.required );
-		}
-	}, [
-		defaultFields.phone.hidden,
-		defaultFields.company.hidden,
-		defaultFields.address_2.hidden,
-		defaultFields.phone.required,
-		defaultFields.company.required,
-		defaultFields.address_2.required,
-	] );
-
-	const setFieldEntity = ( field: string, value: string ) => {
-		if (
-			[ 'phone', 'company', 'address_2' ].includes( field ) &&
-			[ 'optional', 'required', 'hidden' ].includes( value )
-		) {
-			dispatch( coreStore as unknown as string ).editEntityRecord(
-				'root',
-				'site',
-				undefined,
-				{
-					[ `woocommerce_checkout_${ field }_field` ]: value,
-				}
-			);
-		}
-	};
-
 	// This focuses on the block when a certain query param is found. This is used on the link from the task list.
 	const focus = useRef( getQueryArg( window.location.href, 'focus' ) );
 
@@ -167,116 +119,6 @@ export const Edit = ( {
 		[ 'woocommerce/checkout-fields-block', {}, [] ],
 	] as TemplateArray;
 
-	const requiredOptions = [
-		{
-			label: __( 'Optional', 'woocommerce' ),
-			value: 'false',
-		},
-		{
-			label: __( 'Required', 'woocommerce' ),
-			value: 'true',
-		},
-	];
-
-	const addressFieldControls = (): JSX.Element => (
-		<InspectorControls>
-			<PanelBody title={ __( 'Address Fields', 'woocommerce' ) }>
-				<p className="wc-block-checkout__controls-text">
-					{ __(
-						'Show or hide fields in the checkout address forms.',
-						'woocommerce'
-					) }
-				</p>
-				<ToggleControl
-					label={ __( 'Company', 'woocommerce' ) }
-					checked={ ! defaultFields.company.hidden }
-					onChange={ () => {
-						if ( defaultFields.company.hidden ) {
-							setFieldEntity(
-								'company',
-								lastCompanyRequired ? 'required' : 'optional'
-							);
-						} else {
-							setFieldEntity( 'company', 'hidden' );
-						}
-					} }
-				/>
-				{ ! defaultFields.company.hidden && (
-					<RadioControl
-						selected={
-							defaultFields.company.required ? 'true' : 'false'
-						}
-						options={ requiredOptions }
-						onChange={ ( value: string ) => {
-							setFieldEntity(
-								'company',
-								value === 'true' ? 'required' : 'optional'
-							);
-						} }
-						className="components-base-control--nested wc-block-components-require-company-field"
-					/>
-				) }
-				<ToggleControl
-					label={ __( 'Address line 2', 'woocommerce' ) }
-					checked={ ! defaultFields.address_2.hidden }
-					onChange={ () => {
-						if ( defaultFields.address_2.hidden ) {
-							setFieldEntity(
-								'address_2',
-								lastAddress2Required ? 'required' : 'optional'
-							);
-						} else {
-							setFieldEntity( 'address_2', 'hidden' );
-						}
-					} }
-				/>
-				{ ! defaultFields.address_2.hidden && (
-					<RadioControl
-						selected={
-							defaultFields.address_2.required ? 'true' : 'false'
-						}
-						options={ requiredOptions }
-						onChange={ ( value: string ) => {
-							setFieldEntity(
-								'address_2',
-								value === 'true' ? 'required' : 'optional'
-							);
-						} }
-						className="components-base-control--nested wc-block-components-require-address_2-field"
-					/>
-				) }
-				<ToggleControl
-					label={ __( 'Phone', 'woocommerce' ) }
-					checked={ ! defaultFields.phone.hidden }
-					onChange={ () => {
-						if ( defaultFields.phone.hidden ) {
-							setFieldEntity(
-								'phone',
-								lastPhoneRequired ? 'required' : 'optional'
-							);
-						} else {
-							setFieldEntity( 'phone', 'hidden' );
-						}
-					} }
-				/>
-				{ ! defaultFields.phone.hidden && (
-					<RadioControl
-						selected={
-							defaultFields.phone.required ? 'true' : 'false'
-						}
-						options={ requiredOptions }
-						onChange={ ( value: string ) => {
-							setFieldEntity(
-								'phone',
-								value === 'true' ? 'required' : 'optional'
-							);
-						} }
-						className="components-base-control--nested wc-block-components-require-phone-field"
-					/>
-				) }
-			</PanelBody>
-		</InspectorControls>
-	);
 	const blockProps = useBlockPropsWithLocking();
 	return (
 		<div { ...blockProps }>
@@ -301,26 +143,23 @@ export const Edit = ( {
 								'has-dark-controls': hasDarkControls,
 							} ) }
 						>
-							<CheckoutBlockControlsContext.Provider
-								value={ { addressFieldControls } }
+							<CheckoutBlockContext.Provider
+								value={ {
+									showOrderNotes,
+									showPolicyLinks,
+									showReturnToCart,
+									cartPageId,
+									showRateAfterTaxName,
+									showFormStepNumbers,
+									defaultFields,
+								} }
 							>
-								<CheckoutBlockContext.Provider
-									value={ {
-										showOrderNotes,
-										showPolicyLinks,
-										showReturnToCart,
-										cartPageId,
-										showRateAfterTaxName,
-										showFormStepNumbers,
-									} }
-								>
-									<InnerBlocks
-										allowedBlocks={ ALLOWED_BLOCKS }
-										template={ defaultTemplate }
-										templateLock="insert"
-									/>
-								</CheckoutBlockContext.Provider>
-							</CheckoutBlockControlsContext.Provider>
+								<InnerBlocks
+									allowedBlocks={ ALLOWED_BLOCKS }
+									template={ defaultTemplate }
+									templateLock="insert"
+								/>
+							</CheckoutBlockContext.Provider>
 						</SidebarLayout>
 					</CheckoutProvider>
 				</SlotFillProvider>
