+++ plugins/woocommerce-blocks/assets/js/blocks/checkout/inner-blocks/checkout-fields-block/edit.tsx
@@ -10,16 +10,13 @@ import type { TemplateArray } from '@wordpress/blocks';
 /**
  * Internal dependencies
  */
-import {
-	useCheckoutBlockContext,
-	useCheckoutBlockControlsContext,
-} from '../../context';
+import { useCheckoutBlockContext } from '../../context';
 import {
 	useForcedLayout,
 	getAllowedBlocks,
 } from '../../../cart-checkout-shared';
 import './style.scss';
-
+import { AddressFieldControls } from '../../address-field-controls';
 export const Edit = ( {
 	clientId,
 	attributes,
@@ -36,8 +33,6 @@ export const Edit = ( {
 	const allowedBlocks = getAllowedBlocks( innerBlockAreas.CHECKOUT_FIELDS );
 
 	const { showFormStepNumbers } = useCheckoutBlockContext();
-	const { addressFieldControls: Controls } =
-		useCheckoutBlockControlsContext();
 
 	const defaultTemplate = [
 		[ 'woocommerce/checkout-express-payment-block', {}, [] ],
@@ -62,7 +57,7 @@ export const Edit = ( {
 
 	return (
 		<Main { ...blockProps }>
-			<Controls />
+			<AddressFieldControls />
 			<form
 				className={ clsx(
 					'wc-block-components-form wc-block-checkout__form',
